refactor: rename layermeta to layerpopup
