+++ src/components/LayerControl.vue
@@ -1,17 +1,17 @@
 <template>
   <div :class="`layer-control${(active) ? ' layer-control--active' : ''}`">
-    <div class="layer-control__main">
-    <input type="checkbox"
-      class="sr-only layer-control__vis-checkbox"
-      :name="`layer-${id}-vis`"
-      :id="`layer-${id}-vis`"
-      value="zichtbaar"
-      checked
-    >
-    <label
-        class="layer-control__vis-label"
-        :for="`layer-${id}-vis`"
+    <form class="layer-control__main">
+      <input type="checkbox"
+        class="sr-only layer-control__vis-checkbox"
+        :name="`layer-${id}-vis`"
+        :id="`layer-${id}-vis`"
+        value="zichtbaar"
+        checked
       >
+      <label
+          class="layer-control__vis-label"
+          :for="`layer-${id}-vis`"
+        >
       <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 64 64">
         <path fill="none" d="M0 0h64v64H0z"/>
         <path d="M59 28c-3-5-12-16-27-16S8 23 5 28v6c3 5 12 16 27 16s24-11 27-16v-6zm-27-7a6 6 0 1 1 0 12 6 6 0 0 1 0-12zm0 24c-15 0-22-14-22-14s3-6 10-10l-2 4a14 14 0 1 0 28 3c0-5-4-8-9-10 12 2 17 13 17 13s-7 14-22 14z"/>
@@ -26,7 +26,7 @@
           {{ subtitle }}
         </p>
       </div>
-    </div>
+    </form>
     <div class="layer-control__options">
       <select :name="`layer-${id}-trans`">
         <option
@@ -37,24 +37,24 @@
           {{ i * 10 }}% transparantie
         </option>
       </select>
-      <button class="layer-control__info" @click="metaIsOpen=true">
+      <button class="layer-control__info" @click="popupIsOpen=true">
         <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="32" height="32" viewBox="0 0 64 64">
           <path fill="none" d="M0 0h64v64H0z"/>
           <path d="M53.9 14.1c-.4-2-2-3.6-4-4-6-1-16-1-17.9-1-2 0-12 0-17.9 1-2 .4-3.6 2-4 4-1 6-1 16-1 17.9s0 12 1 17.9c.4 2 2 3.6 4 4 6 1 16 1 17.9 1 2 0 12 0 17.9-1 2-.4 3.6-2 4-4 1-6 1-16 1-17.9 0-6 0-12-1-17.9zM35 48h-6.6l.6-14v-8h6v22zm-3-26c-2.2 0-3.5-1.3-3.5-3.5 0-2 1.2-3.5 3.5-3.5 2.2 0 3.5 1.2 3.5 3.5 0 2-1.2 3.5-3.5 3.5z"/>
         </svg>
       </button>
-      <layer-meta v-if="metaIsOpen"
-        :metadata="metadata" @close="metaIsOpen=false" />
+      <layer-popup v-if="popupIsOpen"
+        :metadata="metadata" @close="popupIsOpen=false" />
     </div>
   </div>
 </template>
 
 <script>
-import LayerMeta from '@/components/LayerMeta'
+import LayerPopup from '@/components/LayerPopup'
 
 export default {
   data () {
-    return { metaIsOpen: false }
+    return { popupIsOpen: false }
   },
   props: {
     id: [String, Number],
@@ -67,12 +67,12 @@ export default {
     // Always close meta when switching layers
     active (isActive) {
       if (!isActive) {
-        this.metaIsOpen = false
+        this.popupIsOpen = false
       }
     }
   },
   components: {
-    LayerMeta
+    LayerPopup
   }
 }
 </script>
